Move true up review from cloud sections into hosted customer sections.
