+++ actions/hosted_customer.tsx
@@ -16,6 +16,7 @@ import {getSelfHostedErrors} from 'mattermost-redux/selectors/entities/hosted_cu
 import {StripeSetupIntent, BillingDetails} from 'types/cloud/sku';
 
 import {getConfirmCardSetup} from 'components/payment_form/stripe';
+import { bindClientFunc } from 'mattermost-redux/actions/helpers';
 
 function selfHostedNeedsConfirmation(progress: ValueOf<typeof SelfHostedSignupProgress>): boolean {
     switch (progress) {
@@ -178,4 +179,20 @@ export function retryFailedHostedCustomerFetches() {
 
         return {data: true};
     };
+}
+
+export function getTrueUpReviewBundle(): ActionFunc {
+    return bindClientFunc({
+        clientFunc: Client4.submitTrueUpReview,
+        onSuccess: [HostedCustomerTypes.RECEIVED_TRUE_UP_REVIEW_BUNDLE],
+        onFailure: HostedCustomerTypes.TRUE_UP_REVIEW_FAILED,
+    });
+}
+
+export function getTrueUpReviewStatus(): ActionFunc {
+    return bindClientFunc({
+        clientFunc: Client4.getTrueUpReviewStatus,
+        onSuccess: [HostedCustomerTypes.RECEIVED_TRUE_UP_REVIEW_STATUS],
+        onFailure: HostedCustomerTypes.TRUE_UP_REVIEW_STATUS_FAILED,
+    });
 }
