+++ packages/mattermost-redux/src/action_types/hosted_customer.ts
@@ -11,4 +11,8 @@ export default keyMirror({
     RECEIVED_SELF_HOSTED_PRODUCTS: null,
     RECEIVED_SELF_HOSTED_SIGNUP_PROGRESS: null,
     RECEIVED_SELF_HOSTED_INVOICES: null,
+    RECEIVED_TRUE_UP_REVIEW_BUNDLE: null,
+    TRUE_UP_REVIEW_FAILED: null,
+    RECEIVED_TRUE_UP_REVIEW_STATUS: null,
+    TRUE_UP_REVIEW_STATUS_FAILED: null,
 });
