+++ packages/types/src/cloud.ts
@@ -23,10 +23,6 @@ export type CloudState = {
     selfHostedSignup: {
         progress: ValueOf<typeof SelfHostedSignupProgress>;
     };
-    trueUpReview: {
-        status: TrueUpReviewStatus;
-        profile: TrueUpReviewProfile;
-    };
 }
 
 export type Subscription = {
@@ -218,31 +214,4 @@ export interface CreateSubscriptionRequest {
     add_ons: string[];
     seats: number;
     internal_purchase_order?: string;
-}
-
-export type TrueUpReviewProfile = {
-    server_id: string;
-    server_version: string;
-    server_installation_type: string;
-    license_id: string;
-    licensed_seats: number;
-    license_plan: string;
-    customer_name: string;
-    active_users: number;
-    authentication_features: string[];
-    plugins: TrueUpReviewPlugins;
-    total_incoming_webhooks: number;
-    total_outgoing_webhooks: number;
-}
-
-export type TrueUpReviewPlugins = {
-    total_active_plugins: number;
-    total_inactive_plugins: number;
-    active_plugin_names: string[];
-    inactive_plugin_names: string[];
-}
-
-export type TrueUpReviewStatus = {
-    due_date: number;
-    complete: boolean;
 }
